# frontend/dist/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Kurawal API",
        "version": "1.0.0",
        "description": "Kurawal Backend API Documentation"
    },
    "servers": [
        {
            "url": "/",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/tags": {
            "get": {
                "summary": "Get all tags",
                "description": "Retrieve a list of all available tags",
                "tags": [
                    "Tags"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tags retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "summary": "Create a new tag",
                "description": "Create a new tag",
                "tags": [
                    "Tags"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Web Development"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "web-development",
                                        "description": "Optional; auto-generated if omitted"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag created successfully"
                    },
                    "400": {
                        "description": "Invalid request data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/tags/{id}": {
            "get": {
                "summary": "Get a specific tag",
                "description": "Retrieve a tag by ID",
                "tags": [
                    "Tags"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Tag ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Tag not found"
                    }
                }
            },
            "put": {
                "summary": "Update a tag",
                "description": "Update an existing tag",
                "tags": [
                    "Tags"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Tag ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Web Development"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "web-development",
                                        "description": "Optional; auto-generated if omitted"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tag updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Tag not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a tag",
                "description": "Delete a tag by ID",
                "tags": [
                    "Tags"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "Tag ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Tag not found"
                    }
                }
            }
        },
...
    },
...
}